third_party/WebKit/LayoutTests/http/tests/devtools/elements/highlight/highlight-dom-updates-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests DOM update highlights in the DOM tree.


Running: testDumpInitial
========= Original ========
<No highlights>

Running: testSetAttributeOtherValue
//*[@id="attrTest"]/@attrfoo bar

Running: testSetAttributeEmpty
//*[@id="attrTest"]/@attrfoo (empty)

Running: testAddAttribute
//*[@id="attrTest"]/@attrbar newBar

Running: testRemoveAttribute
//*[@id="attrTest"]

Running: testAppendChildToEmpty
//*[@id="childTest"]

Running: testAppendChildToExpanded
    //*[@id="child2"]

Running: testRemoveChild1
//*[@id="childTest"]

Running: testRemoveChild2
//*[@id="childTest"]

Running: testSetTextContent
//*[@id="textTest"]
//*[@id="textTest"]/text() "Text"

Running: testSetTextNodeTextContent
//*[@id="textTest"]
//*[@id="textTest"]/text() "NewText"

Running: testRemoveInlineTextNode
//*[@id="textTest"]

Running: testSetTextContentWithEmptyText
//*[@id="textTest"]
//*[@id="textTest"]/text() "Text"

Running: testClearTextNodeTextContent
//*[@id="textTest"]
//*[@id="textTest"]/text() ""

Running: testAppendChildWhenHidden
<No highlights>
